dp done

+++ server/src/main/java/pt/unl/fct/pds/proj1server/model/CountResponse.java
@@ -1,24 +1,32 @@
 package pt.unl.fct.pds.proj1server.model;
 
-import jakarta.persistence.*;
-
 public class CountResponse {
     private String attribute;
-    private int value;
+    private long value;
 
     public CountResponse() {
     }
 
     public CountResponse(
-        String attribute,
-        int value) {
+            String attribute,
+            long value) {
         this.attribute = attribute;
         this.value = value;
     }
-    
-    public String getAttribute() {return attribute;}
-    public int getValue() {return value;}
-    
-    public void setAttribute(String attribute) {this.attribute = attribute;}
-    public void setValue(int value) {this.value = value;}
+
+    public String getAttribute() {
+        return attribute;
+    }
+
+    public long getValue() {
+        return value;
+    }
+
+    public void setAttribute(String attribute) {
+        this.attribute = attribute;
+    }
+
+    public void setValue(int value) {
+        this.value = value;
+    }
 }
